delete tests module and add test folder
